# Conflicts:
#	src/main/java/DWS/practica_dws/model/Product.java
#	src/main/resources/templates/showProduct.html

+++ src/main/java/DWS/practica_dws/controller/ProductController.java
@@ -1,9 +1,11 @@
 package DWS.practica_dws.controller;
 
+import DWS.practica_dws.model.Comment;
 import DWS.practica_dws.model.Product;
 import DWS.practica_dws.service.ImageService;
 import DWS.practica_dws.service.ProductsService;
 import DWS.practica_dws.service.UserSession;
+import jakarta.jws.WebParam;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -98,15 +100,20 @@ public class ProductController {
         model.addAttribute("price", prod1.getPrice());
         model.addAttribute("description", prod1.getDescription());
         model.addAttribute("image", "/images/products/" + id + ".jpeg"); // Ruta de la imagen
+        model.addAttribute("commentList", prod1.getComments());
 
         return "showProduct";
     }
 
-    @GetMapping("/product/{id}/delete")
+    @PostMapping("/product/{id}/delete")
     public String deleteProduct(Model model, @PathVariable long id) {
-        productsService.deleteProduct(id);
+        Product p = productsService.getProduct(id);
 
-        model.addAttribute("allProducts", productsService.getAll());
+        if(p!=null){
+            productsService.deleteProduct(id);
+            model.addAttribute("product", p.getName());
+            model.addAttribute("exist", true);
+        } else model.addAttribute("exist", false);
 
         return "deletedProduct";
     }
@@ -116,6 +123,22 @@ public class ProductController {
         productsService.removeProductFromCart(id, userSession);
         model.addAttribute("products", userSession.userProducts());
         return "shoppingCart";
+    }
+
+
+    @PostMapping("/product/{id}/newComment")
+    public String newComment(Model model, @PathVariable long id, @RequestParam String userName,
+                             @RequestParam int score, @RequestParam String comment){
+        Product p = productsService.getProduct(id);
+        if(p!=null){
+            p.addComment(new Comment(userName, score, comment));
+            model.addAttribute("name", p.getName());
+            model.addAttribute("price", p.getPrice());
+            model.addAttribute("description", p.getDescription());
+            model.addAttribute("image", "/images/products/" + id + ".jpeg"); // Ruta de la imagen
+            model.addAttribute("commentList", p.getComments());
+        }
+        return "showProduct";
     }
 
 }
